requirements.txt
black==23.11.0
Django==4.2.7
flake8==6.1.0
mysqlclient==2.2.0
Pillow==10.1.0
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.4.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
requests==2.31.0
redis==5.0.1
celery==5.3.6
django-celery-beat==2.5.0
